First webpack version
