+++ .gitignore
@@ -2,6 +2,7 @@
 startdev.bat
 myinfo.code-workspace
 build
+dist
 public/myinfo.js
 public/lp_thebrain.js
 public/dp_note.js
