+++ package.json
@@ -4,15 +4,17 @@
   "description": "MyInfo: Just another mindmap application.",
   "main": "myinfoServer.js",
   "scripts": {
-    "build": "run-s clean build_myinfo deploy_myinfo build_lp_thebrain deploy_lp_thebrain build_dp_note deploy_dp_note",
-    "clean": "del build\\*.js",
+    "build": "webpack",
+    "webpack-dev-server": "webpack-dev-server",
+    "start": "start node myinfoServer.js",
+    "build-old": "run-s clean build_myinfo deploy_myinfo build_lp_thebrain deploy_lp_thebrain build_dp_note deploy_dp_note",
+    "clean-old": "del build\\*.js",
     "build_myinfo": "rollup -c myinfo.rollup.config.js",
     "deploy_myinfo": "copy /Y build\\myinfo.js public\\myinfo.js",
     "build_lp_thebrain": "rollup -c lp_thebrain.rollup.config.js",
     "deploy_lp_thebrain": "copy /Y build\\lp_thebrain.js public\\lp_thebrain.js",
     "build_dp_note": "rollup -c dp_note.rollup.config.js",
-    "deploy_dp_note": "copy /Y build\\dp_note.js public\\dp_note.js",
-    "start": "start node myinfoServer.js"
+    "deploy_dp_note": "copy /Y build\\dp_note.js public\\dp_note.js"
   },
   "repository": {
     "type": "git",
@@ -22,7 +24,7 @@
     "mindmap",
     "thebrain"
   ],
-  "author": "\\dev\\null",
+  "author": "/dev/null",
   "contributors": [
     {
       "name": "Szil",
@@ -35,11 +37,15 @@
   },
   "homepage": "https://github.com/perdevpernull/myinfo#readme",
   "dependencies": {
-    "express": "^4.16.2"
+    "express": "^4.16.3"
   },
   "devDependencies": {
-    "eslint": "^4.13.0",
-    "npm-run-all": "^4.1.2",
-    "rollup": "^0.52.1"
+    "clean-webpack-plugin": "^0.1.19",
+    "eslint": "^4.19.1",
+    "file-loader": "^1.1.11",
+    "html-webpack-plugin": "^3.2.0",
+    "webpack": "^4.16.3",
+    "webpack-cli": "^3.1.0",
+    "webpack-dev-middleware": "^3.1.3"
   }
 }
