+++ src/main_myinfo.js
@@ -1,0 +1,8 @@
+import {MyInfo} from "./myinfo/myinfo";
+
+// Start the program
+global.myinfo = new MyInfo();
+global.myinfo.run("#myinfo");
+
+// MyInfo API
+//export {refreshHome, openDataset, registerDataPlugin, registerLayoutPlugin} from "./myinfo/main";
