--- src/myinfo.js
@@ -1,7 +1,0 @@
-import {run} from "./myinfo/main";
-
-// Start the program
-run("#myinfo");
-
-// MyInfo API
-export {refreshHome, openDataset, registerDataPlugin, registerLayoutPlugin} from "./myinfo/main";
