+++ src/myinfo/dataset-ID0.json
@@ -1,0 +1,78 @@
+{
+	"nextNodeID": 7,
+	"nodes": {
+		"ID0": {
+			"ID": 0,
+			"text": "Node 0",
+			"links": [
+				{"toID": 2, "type": "parent", "tags": [], "deleted": false},
+				{"toID": 1, "type": "child", "tags": [], "deleted": false}
+			],
+			"tags": [],
+			"deleted": false
+		},
+		"ID1": {
+			"ID": 1,
+			"text": "Node 1",
+			"links": [
+				{"toID": 0, "type": "parent", "tags": [], "deleted": false},
+				{"toID": 2, "type": "child", "tags": [], "deleted": false}
+			],
+			"tags": [],
+			"deleted": false
+		},
+		"ID2": {
+			"ID": 2,
+			"text": "Node 2",
+			"links": [
+				{"toID": 1, "type": "parent", "tags": [], "deleted": false},
+				{"toID": 0, "type": "child", "tags": [], "deleted": false}
+			],
+			"tags": [],
+			"deleted": false
+		},
+		"ID3": {
+			"ID": 3,
+			"text": "Node 3",
+			"links": [
+				{"toID": 4, "type": "child", "tags": [], "deleted": false},
+				{"toID": 5, "type": "child", "tags": [], "deleted": false}
+			],
+			"tags": [],
+			"deleted": false
+		},
+		"ID4": {
+			"ID": 4,
+			"text": "Node 4",
+			"links": [
+				{"toID": 3, "type": "parent", "tags": [], "deleted": false}
+			],
+			"tags": [],
+			"deleted": false
+		},
+		"ID5": {
+			"ID": 5,
+			"text": "Node 5",
+			"links": [
+				{"toID": 3, "type": "parent", "tags": [], "deleted": false},
+				{"toID": 6, "type": "child", "tags": [], "deleted": false}
+			],
+			"tags": [],
+			"deleted": false
+		},
+		"ID6": {
+			"ID": 6,
+			"text": "Node 6",
+			"links": [
+				{"toID": 5, "type": "parent", "tags": [], "deleted": false}
+			],
+			"tags": [],
+			"deleted": false
+		}
+	},
+	"tags": {},
+	"rootNodes": {
+		"ID3": 3
+	},
+	"pluginData": {}
+}
