+++ src/myinfo/settings.json
@@ -1,0 +1,3 @@
+{
+	"test": 0
+}
