+++ src/myinfo/userdata-ID0.json
@@ -1,0 +1,3 @@
+{
+	"test": 0
+}
