+++ src/template.myinfo.html
@@ -1,0 +1,49 @@
+<!DOCTYPE html>
+<html lang="en">
+	<head>
+		<!-- Required meta tags -->
+		<meta charset="utf-8">
+		<meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+
+		<!-- Meta -->
+		<title> <%= htmlWebpackPlugin.options.title %> </title>
+		<meta name="description" content="">
+		<meta name="author" content="">
+
+		<!-- Favicons -->
+		<!--link rel="apple-touch-icon" href="/apple-touch-icon.png"-->
+		<!--link rel="icon" href="favicon.ico"-->
+
+
+		<!-- Bootstrap CSS -->
+		<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta/css/bootstrap.min.css" />
+
+		<!-- Custom styles for this template -->
+		<style>
+		</style>
+
+	</head>
+	<body>
+		<div id="myinfo" class="container-fluid" style="height: 100vh;">
+			<!-- Navigációs menü -->
+
+			<!--div id="sidebar">Sidebar</div-->
+
+		</div>
+
+		<!-- jQuery first, then Popper.js, then Bootstrap JS -->
+		<script src="https://code.jquery.com/jquery-3.2.1.js"></script>
+		<script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.11.0/umd/popper.min.js"></script>
+		<script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.0.0-beta/js/bootstrap.min.js"></script>
+
+		<!-- Select2 -->
+		<script src="https://cdnjs.cloudflare.com/ajax/libs/select2/3.4.4/select2.min.js"></script>
+		<link rel="stylesheet" type="text/css" href="https://cdnjs.cloudflare.com/ajax/libs/select2/3.4.4/select2.css" />
+
+		<!-- D3.js -->
+		<script src="https://cdnjs.cloudflare.com/ajax/libs/d3/4.9.1/d3.min.js"></script>
+
+		<!-- MyInfo JavaScript -->
+		<!--script src="myinfo.js"></script-->
+	</body>
+</html>
